refactor: move exceptions into a dedicated module
Exceptions require some boilerplate to render nicely and obstruct the
flow of the definition of the main functionality of the package.
Extracting them into a separate module decreases the amount of code
necessary to grasp the core concepts of the package in both the package
itself as well as its tests.
