+++ src/Kroki.jl
@@ -12,7 +12,6 @@ module Kroki
 using Base64: base64encode
 using CodecZlib: ZlibCompressor, transcode
 using HTTP: request
-using HTTP.ExceptionRequest: StatusError
 using Reexport: @reexport
 
 include("./kroki/documentation.jl")
@@ -62,6 +61,9 @@ of diagram.
 """
 Diagram(type::Symbol, specification::AbstractString) = Diagram(specification, type)
 
+include("./kroki/exceptions.jl")
+using .Exceptions: DiagramPathOrSpecificationError, RenderError
+
 """
 Constructs a [`Diagram`](@ref) from the `specification` for a specific `type`
 of diagram, or loads the `specification` from the provided `path`.
@@ -88,80 +90,6 @@ function Diagram(
 end
 
 """
-An `Exception` to be thrown when the `path` and `specification` keyword
-arguments to [`Diagram`](@ref) are not specified mutually exclusive.
-"""
-struct DiagramPathOrSpecificationError <: Exception
-  path::Maybe{AbstractString}
-  specification::Maybe{AbstractString}
-end
-
-function Base.showerror(io::IO, error::DiagramPathOrSpecificationError)
-  not_specified = "<not specified>"
-
-  path_description = isnothing(error.path) ? not_specified : error.path
-  specification_description =
-    isnothing(error.specification) ? not_specified : error.specification
-
-  message = """
-            Either `path` or `specification` should be specified:
-              * `path`: '$(path_description)'
-              * `specification`: '$(specification_description)'
-            """
-
-  print(io, message)
-end
-
-"""
-An `Exception` to be thrown when a [`Diagram`](@ref) representing an invalid
-specification is passed to [`render`](@ref).
-"""
-struct InvalidDiagramSpecificationError <: Exception
-  error::String
-  cause::Diagram
-end
-
-Base.showerror(io::IO, error::InvalidDiagramSpecificationError) = print(
-  io,
-  """
-  $(RenderErrorHeader(error))
-
-  This is (likely) caused by an invalid diagram specification.
-  """,
-)
-
-"""
-An `Exception` to be thrown when a [`Diagram`](@ref) is [`render`](@ref)ed to
-an unsupported or invalid output format.
-"""
-struct InvalidOutputFormatError <: Exception
-  error::String
-  cause::Diagram
-end
-
-Base.showerror(io::IO, error::InvalidOutputFormatError) = print(
-  io,
-  """
-  $(RenderErrorHeader(error))
-
-  This is (likely) caused by an invalid or unknown output format.
-  """,
-)
-
-# Helper function to render common headers when showing render errors
-function RenderErrorHeader(
-  error::Union{InvalidDiagramSpecificationError, InvalidOutputFormatError},
-)
-  """
-  The Kroki service responded with:
-  $(error.error)
-
-  In response to a '$(error.cause.type)' diagram with the specification:
-  $(error.cause.specification)
-  """
-end
-
-"""
 Compresses a [`Diagram`](@ref)'s `specification` using
 [zlib](https://zlib.net), turning the resulting bytes into a URL-safe Base64
 encoded payload (i.e. replacing `+` by `-` and `/` by `_`) to be used in
@@ -176,33 +104,17 @@ UriSafeBase64Payload(diagram::Diagram) = foldl(
   init = base64encode(transcode(ZlibCompressor, diagram.specification)),
 )
 
-# Rewrites generic `HTTP.ExceptionRequest.StatusError`s into more specific
-# errors based on Kroki's response if possible
-function RenderError(diagram::Diagram, exception::StatusError)
-  # Both errors related to invalid diagram specifications and invalid or
-  # unsupported output formats are denoted by 400 responses, so further
-  # processing of the response is necessary
-  service_response = String(exception.response.body)
-
-  if occursin("Unsupported output format", service_response)
-    InvalidOutputFormatError(service_response, diagram)
-  elseif occursin("Syntax Error", service_response)
-    InvalidDiagramSpecificationError(service_response, diagram)
-  else
-    exception
-  end
-end
-RenderError(::Diagram, exception::Exception) = exception
-
 """
 Renders a [`Diagram`](@ref) through a Kroki service to the specified output
 format.
 
 If the Kroki service responds with an error, throws an
-[`InvalidDiagramSpecificationError`](@ref) or
-[`InvalidOutputFormatError`](@ref) if a know type of error occurs. Other errors
-(e.g. `HTTP.ExceptionRequest.StatusError` for connection errors) are propagated
-if they occur.
+[`InvalidDiagramSpecificationError`](@ref
+Kroki.Exceptions.InvalidDiagramSpecificationError) or
+[`InvalidOutputFormatError`](@ref Kroki.Exceptions.InvalidOutputFormatError) if
+a know type of error occurs. Other errors (e.g.
+`HTTP.ExceptionRequest.StatusError` for connection errors) are propagated if
+they occur.
 
 _SVG output is supported for all [`Diagram`](@ref) types_. See [Kroki's
 website](https://kroki.io/#support) for an overview of other supported output
