+++ test/runtests.jl
@@ -2,36 +2,8 @@ module KrokiTest
 
 using Test: @testset, @test, @test_nowarn, @test_throws
 
-using Kroki:
-  Diagram,
-  DiagramPathOrSpecificationError,
-  InvalidDiagramSpecificationError,
-  InvalidOutputFormatError,
-  StatusError, # Imported from HTTP through Kroki
-  render
-using Kroki.Service: setEndpoint!
-
-testRenderError(
-  title::AbstractString,
-  diagram_type::Symbol,
-  content::AbstractString,
-  output_format::AbstractString,
-  expected_error_type::DataType,
-) = @testset "$(title)" begin
-  diagram = Diagram(diagram_type, content)
-
-  @test_throws(expected_error_type, render(diagram, output_format))
-
-  @testset "rendering" begin
-    service_response = "$(title) response"
-
-    rendered_error = sprint(showerror, expected_error_type(service_response, diagram))
-
-    @test occursin("Kroki service responded with:", rendered_error)
-    @test occursin(content, rendered_error)
-    @test occursin(service_response, rendered_error)
-  end
-end
+using Kroki: Diagram, render
+using Kroki.Exceptions: InvalidOutputFormatError
 
 function testShowMethodRenders(
   diagram::Diagram,
@@ -42,8 +14,8 @@ function testShowMethodRenders(
 end
 
 @testset "Kroki" begin
-  @testset "`Diagram` instantiation providing" begin
-    @testset "`path` loads the file as the `specification" begin
+  @testset "`Diagram` instantiation" begin
+    @testset "providing `path` loads the file as the `specification" begin
       diagram_path = joinpath(@__DIR__, "assets", "plantuml-example.puml")
       expected_specification = read(diagram_path, String)
 
@@ -52,39 +24,12 @@ end
       @test diagram.specification === expected_specification
     end
 
-    @testset "`specification` stores it" begin
+    @testset "providing `specification` stores it" begin
       expected_specification = "A -> B: C"
 
       diagram = Diagram(:plantuml; specification = expected_specification)
 
       @test diagram.specification === expected_specification
-    end
-
-    @testset "invalid `path`/`specification` combinations errors" begin
-      @testset "specifying both" begin
-        @test_throws(
-          DiagramPathOrSpecificationError,
-          Diagram(:mermaid; path = tempname(), specification = "A -> B: C")
-        )
-      end
-
-      @testset "specifying neither" begin
-        @test_throws(DiagramPathOrSpecificationError, Diagram(:svgbob))
-      end
-
-      @testset "rendering" begin
-        expected_specification = "X -> Y: Z"
-
-        rendered_error =
-          sprint(showerror, DiagramPathOrSpecificationError(nothing, "X -> Y: Z"))
-
-        @test startswith(
-          rendered_error,
-          "Either `path` or `specification` should be specified:",
-        )
-        @test occursin("* `path`: '<not specified>'", rendered_error)
-        @test occursin("* `specification`: '$(expected_specification)'", rendered_error)
-      end
     end
   end
 
@@ -129,63 +74,6 @@ end
       # Some renderers (e.g. Graphviz) include additional whitespace/newlines
       # after the render, these should be ignored when matching
       @test endswith(rendered, r"</svg>\s?")
-    end
-
-    @testset "`RenderError`" begin
-      # `RenderError`s are thrown from the `render` function whenever an error
-      # occurs. Some of these benefit from rewrites into more descriptive
-      # errors
-      @testset "transforms rewritable `HTTP.ExceptionRequest.StatusError`s" begin
-        # Errors in the Kroki service are thrown as
-        # `HTTP.ExceptionRequest.StatusError` and should be rewritten in more
-        # descriptive errors
-        testRenderError(
-          "invalid diagram specification",
-          :PlantUML,
-          # A missing `>` and message cause the specification to be invalid
-          "Julia - Kroki:",
-          "svg",
-          InvalidDiagramSpecificationError,
-        )
-
-        testRenderError(
-          "invalid output format",
-          :Mermaid,
-          # The Mermaid renderer does not support PNG output
-          "graph TD; A --> B;",
-          "jpeg",
-          InvalidOutputFormatError,
-        )
-      end
-
-      @testset "passes unknown `HTTP.ExceptionRequest.StatusError`s as-is" begin
-        # Any HTTP related errors that are not due to rendering errors in the
-        # Kroki service (e.g. unknown endpoints), should be thrown from
-        # `RenderError` as-is
-        @test_throws(StatusError, render(Diagram(:non_existent_diagram_type, ""), "svg"))
-      end
-
-      @testset "passes other errors as-is" begin
-        # Non-`StatusError`s (e.g. `IOError`s due to incorrect hostnames should
-        # be thrown/returned as-is
-        expected_service_host = "http://localhost:1"
-        expected_diagram_type = :plantuml
-        setEndpoint!(expected_service_host)
-
-        try
-          render(Diagram(expected_diagram_type, "A -> B: C"), "svg")
-        catch exception
-          rendered_buffer = sprint(showerror, exception)
-
-          @test occursin("ECONNREFUSED", rendered_buffer)
-          @test occursin(
-            "$(expected_service_host)/$(expected_diagram_type)",
-            rendered_buffer,
-          )
-        end
-
-        setEndpoint!()
-      end
     end
   end
